moved bottom sheet from CreateEventSheet directly into MapsPage to fix issue with refs
