+++ components/elements/CreateEventSheet.js
@@ -1,7 +1,6 @@
 import { AntDesign } from "@expo/vector-icons";
-import BottomSheet from "@gorhom/bottom-sheet";
 import DateTimePicker from "@react-native-community/datetimepicker";
-import React, { useMemo, useRef, useState } from "react";
+import React, { useState } from "react";
 import { Button, View, Text, Platform } from "react-native";
 import { GooglePlacesAutocomplete } from "react-native-google-places-autocomplete";
 
@@ -41,8 +40,6 @@ const GooglePlacesInput = (props) => {
 };
 
 const CreateEventSheet = (props) => {
-  const createEventBottomSheetRef = useRef(null);
-  const snapPoints = useMemo(() => ["96%"], []);
   const headerText = "Details";
   const [dateOfEvent, setDateOfEvent] = useState(new Date());
   const [timeOfEvent, setTimeOfEvent] = useState(new Date());
@@ -89,146 +86,137 @@ const CreateEventSheet = (props) => {
   };
 
   return (
-    <BottomSheet
-      ref={createEventBottomSheetRef}
-      snapPoints={snapPoints}
-      enablePanDownToClose
-      onClose={props.onClose}
-      onChange={props.onChange}
-    >
-      <View style={[Styles.page, { gap: 15 }]}>
-        <SheetHeader sheetHeaderText={headerText} />
-
-        <_Input
-          labelText="Event Name"
-          borderColor="primary1_100"
-          placeholder="Event name"
-          placeholderTextColor="primary1_030"
-          onChangeText={setEventName}
-        />
-        <View
-          style={{
-            flexDirection: "row",
-            justifyContent: "space-between",
-            marginBottom: 30,
-          }}
-        >
-          <GooglePlacesInput onPress={onPressLocation} />
-        </View>
-        <View
-          style={{
-            gap: 15,
-            flexDirection: "row",
-          }}
-        >
-          <View>
-            <Text style={[Styles.inputLabel]}>Date</Text>
-            <View
-              style={[
-                Styles.d2Box,
-                {
-                  flexDirection: "row",
-                  borderRadius: 10,
-                  fontSize: 17,
-                  borderColor: Colors.primary1_100,
-                  justifyContent: "flex-start",
-                  alignItems: "center",
-                  padding: 12,
-                  height: 50,
-                  width: 160,
-                },
-              ]}
-            >
-              <AntDesign
-                name="calendar"
-                size={18}
-                style={{
-                  color: Colors.primary1_100,
-                }}
-              />
-              {/* Added platform code because there's a bug in Android 
-          where the date picker won't hide */}
-              {Platform.OS === "ios" ? (
-                <DateTimePicker
-                  value={dateOfEvent}
-                  onChange={onChangeDateOfEvent}
-                />
-              ) : showDate ? (
-                <DateTimePicker
-                  value={dateOfEvent}
-                  onChange={onChangeDateOfEvent}
-                />
-              ) : (
-                <Button
-                  title={dateOfEvent.toLocaleDateString()}
-                  onPress={() => setShowDate(true)}
-                />
-              )}
-            </View>
-          </View>
-          <View>
-            <Text style={[Styles.inputLabel]}>Time</Text>
-            <View
-              style={[
-                Styles.d2Box,
-                {
-                  flexDirection: "row",
-                  borderRadius: 10,
-                  fontSize: 17,
-                  borderColor: Colors.primary1_100,
-                  justifyContent: "flex-start",
-                  alignItems: "center",
-                  padding: 12,
-                  height: 50,
-                  width: 160,
-                },
-              ]}
-            >
-              <AntDesign
-                name="clockcircleo"
-                size={18}
-                style={{
-                  color: Colors.primary1_100,
-                }}
-              />
-              {/* Added platform code because there's a bug in Android 
-          where the date picker won't hide */}
-              {Platform.OS === "ios" ? (
-                <DateTimePicker
-                  mode="time"
-                  value={timeOfEvent}
-                  onChange={onChangeTimeOfEvent}
-                />
-              ) : showTime ? (
-                <DateTimePicker
-                  mode="time"
-                  value={timeOfEvent}
-                  onChange={onChangeTimeOfEvent}
-                />
-              ) : (
-                <Button
-                  title={timeOfEvent.toLocaleDateString()}
-                  onPress={() => setShowTime(true)}
-                />
-              )}
-            </View>
-          </View>
-        </View>
-
-        <_Input
-          labelText="Description"
-          borderColor="primary1_100"
-          placeholder="Description"
-          placeholderTextColor="primary1_030"
-          subtextText={`${eventDescription.length}/100`}
-          subtextColor="primary1_030"
-          subtextAlignSelf="flex-end"
-          onChangeText={setEventDescription}
-        />
-        <_Button text="Create Event" action={handleCreateEvent} />
-        <_Button text="Cancel" type="secondary" />
+    <View style={[Styles.page, { gap: 15 }]}>
+      <SheetHeader sheetHeaderText={headerText} />
+      <_Input
+        labelText="Event Name"
+        borderColor="primary1_100"
+        placeholder="Event name"
+        placeholderTextColor="primary1_030"
+        onChangeText={setEventName}
+      />
+      <View
+        style={{
+          flexDirection: "row",
+          justifyContent: "space-between",
+          marginBottom: 30,
+        }}
+      >
+        <GooglePlacesInput onPress={onPressLocation} />
       </View>
-    </BottomSheet>
+      <View
+        style={{
+          gap: 15,
+          flexDirection: "row",
+        }}
+      >
+        <View>
+          <Text style={[Styles.inputLabel]}>Date</Text>
+          <View
+            style={[
+              Styles.d2Box,
+              {
+                flexDirection: "row",
+                borderRadius: 10,
+                fontSize: 17,
+                borderColor: Colors.primary1_100,
+                justifyContent: "flex-start",
+                alignItems: "center",
+                padding: 12,
+                height: 50,
+                width: 160,
+              },
+            ]}
+          >
+            <AntDesign
+              name="calendar"
+              size={18}
+              style={{
+                color: Colors.primary1_100,
+              }}
+            />
+            {/* Added platform code because there's a bug in Android 
+          where the date picker won't hide */}
+            {Platform.OS === "ios" ? (
+              <DateTimePicker
+                value={dateOfEvent}
+                onChange={onChangeDateOfEvent}
+              />
+            ) : showDate ? (
+              <DateTimePicker
+                value={dateOfEvent}
+                onChange={onChangeDateOfEvent}
+              />
+            ) : (
+              <Button
+                title={dateOfEvent.toLocaleDateString()}
+                onPress={() => setShowDate(true)}
+              />
+            )}
+          </View>
+        </View>
+        <View>
+          <Text style={[Styles.inputLabel]}>Time</Text>
+          <View
+            style={[
+              Styles.d2Box,
+              {
+                flexDirection: "row",
+                borderRadius: 10,
+                fontSize: 17,
+                borderColor: Colors.primary1_100,
+                justifyContent: "flex-start",
+                alignItems: "center",
+                padding: 12,
+                height: 50,
+                width: 160,
+              },
+            ]}
+          >
+            <AntDesign
+              name="clockcircleo"
+              size={18}
+              style={{
+                color: Colors.primary1_100,
+              }}
+            />
+            {/* Added platform code because there's a bug in Android 
+          where the date picker won't hide */}
+            {Platform.OS === "ios" ? (
+              <DateTimePicker
+                mode="time"
+                value={timeOfEvent}
+                onChange={onChangeTimeOfEvent}
+              />
+            ) : showTime ? (
+              <DateTimePicker
+                mode="time"
+                value={timeOfEvent}
+                onChange={onChangeTimeOfEvent}
+              />
+            ) : (
+              <Button
+                title={timeOfEvent.toLocaleDateString()}
+                onPress={() => setShowTime(true)}
+              />
+            )}
+          </View>
+        </View>
+      </View>
+
+      <_Input
+        labelText="Description"
+        borderColor="primary1_100"
+        placeholder="Description"
+        placeholderTextColor="primary1_030"
+        subtextText={`${eventDescription.length}/100`}
+        subtextColor="primary1_030"
+        subtextAlignSelf="flex-end"
+        onChangeText={setEventDescription}
+      />
+      <_Button text="Create Event" action={handleCreateEvent} />
+      <_Button text="Cancel" type="secondary" />
+    </View>
   );
 };
